support safetensors model training
